+++ urc_util/include/urc_util/ethernet_socket.hpp
@@ -15,7 +15,16 @@
 class EthernetSocket
 {
 public:
+  /**
+  Constructor opens the socket using UDP and binds it to an
+  endpoint. This endpoint is constructed using the broadcast ip address and
+  specified port.
+
+  @param[in] port port to communicate through
+  @throw std::runtime_error is socket couldn't open or connect
+  */
   EthernetSocket(int port);
+  
   /**
   Constructor opens the socket using UDP and binds it to an
   endpoint. This endpoint is constructed using the specified ip address and
