+++ urc_util/src/ethernet_socket.cpp
@@ -12,8 +12,7 @@ EthernetSocket::EthernetSocket(std::string ip_addr, int port)
 {
   // resolve all possible endpoints
   ip::udp::resolver resolver(io_service_);
-  // ip::udp::resolver::query query(ip_addr, std::to_string(port));
-  ip::udp::resolver::query query(ip::udp::v4(), std::to_string(port));
+  ip::udp::resolver::query query(ip_addr, std::to_string(port));
   ip::udp::resolver::iterator endpoint_iterator = resolver.resolve(query);
 
   // look through endpoints and hit socket's connect() member function until
@@ -30,13 +29,8 @@ EthernetSocket::~EthernetSocket()
 
 void EthernetSocket::sendMessage(char * message, size_t len)
 {
-  // boost::system::error_code error;
-  // sock_->send(boost::asio::buffer(message, len), 0, error);
-
-  char test[] = "hello back";
-
   ip::udp::endpoint senderEndpoint(ip::address_v4::broadcast(), 8443);
-  this->sock_->send_to(boost::asio::buffer(test), senderEndpoint);
+  this->sock_->send_to(boost::asio::buffer(message, len), senderEndpoint);
 }
 
 size_t EthernetSocket::readMessage(unsigned char (& buffer)[256])
